+++ src/content/lib/request-processor.redirects.js
@@ -114,7 +114,11 @@ RequestProcessor = (function(self) {
     {
       let result = PolicyManager.checkRequestAgainstUserRules(originURIObj,
           destURIObj);
-      // For now, we always give priority to deny rules.
+      // For user rules, use the default policy if both types of rule match.
+      if (result.denyRulesExist() && result.allowRulesExist()) {
+        result.isAllowed = Prefs.isDefaultAllow();
+        return result;
+      }
       if (result.denyRulesExist()) {
         result.isAllowed = false;
         return result;